Added a logging system and more init stuff
Added a logger as part of BrickwareUtils. It will cause a crash if you
don't have a Logs directory available in the TestBuild.

Init system should calmly shutdown now if it encounters a problem. It
won't log properly yet but that's next. If the logger fails it will print
to the console and pause.